Move main() for command line `gp_upgrade` into its own directory
Going forward, we'll build three binaries:
 - gp_upgrade
 - gp_upgrade_hub
 - gp_upgrade_agent

`gp_upgrade` is in the `cli/` directory because it's the CLI client; the
others are intended to be daemonized processes

Regarding package declarations: each of the three binaries has its own
main() function, and thus needs to be in the `package main`. We want to
keep each main.go in a separate directory for clarity. Implementation
files for each of the binaries should not be in a `package main` -- they
should be in a distinct, isolated package. A consequence is that all
other *.go files for the implementations of each of the other three
binaries cannot be directly in the top level of `cli/` or `hub/` or
`agent/` directories.

Many of the integration tests will need to change for the new
architecture of binaries / processes, but they're easily made to pass
for now, so we kept them green rather than commenting them out.

Signed-off-by: Shoaib Lari <slari@pivotal.io>